requirements.txt
# Discord 봇 라이브러리
discord.py>=2.3.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 비동기 HTTP 클라이언트
aiohttp>=3.9.0

# PostgreSQL 비동기 드라이버
asyncpg>=0.29.0

# HTML 파싱
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Brotli 압축 해제 지원
brotli>=1.1.0

# 날짜/시간 파싱
python-dateutil>=2.8.0